Record timezone offset when saving event schedules so repeat times can be calculated in the originating time zone.
For example, if the admin is across the international date line from the realm
time zone, and he elects to repeat an event on Tuesdays, the event may be on
Mondays in realm time.

Any "day of week" calculation must be done in the timezone of the person who
created the schedule, since all times are in local time in the UI.

This technique probably fails with daylight saving time. It may be necessary
to present admins with a list of timezones and allow them to chose their
timezone when creating event schedules, in order to know if the admins live in
a region that observes daylight saving time.

* .../gotefarm/client/JSEventSchedule.java
  (JSEventSchedule): Add timezone_offset member.

* .../gotefarm/client/ScheduleEditor.java
  (Schedule): Populate timezone_offset based on what a Date object thinks
  its timezone offset is. This may or may not deal with daylight saving time
  properly.
  (ScheduleEditor): When editing an existing schedule, update the
  timezone_offset to that of the person changing the schedule.

* .../gotefarm/server/dao/GoteFarmJdbcDao.scala
  (JSEventScheduleMapper): Fetch the timezone_offset column.
  (generateTables): Add timezone_offset column to eventsched table.
  (saveEventSchedule): Populate timezone_offset column.